+++ src/main/java/pl/gson/CurrencyHistoryDto.java
@@ -2,12 +2,14 @@ package pl.gson;
 
 import com.google.gson.internal.LinkedTreeMap;
 
+import java.math.BigDecimal;
+
 public class CurrencyHistoryDto {
     private String base;
-    private Double value;
+    private BigDecimal value;
     private String currency;
     private String date;
-    private LinkedTreeMap<String, LinkedTreeMap<String, Double>> rates;
+    private LinkedTreeMap<String, LinkedTreeMap<String, BigDecimal>> rates;
 
     public String getBase() {
         return base;
@@ -17,11 +19,11 @@ public class CurrencyHistoryDto {
         this.base = base;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
 
@@ -41,11 +43,11 @@ public class CurrencyHistoryDto {
         this.date = date;
     }
 
-    public LinkedTreeMap<String, LinkedTreeMap<String, Double>> getRates() {
+    public LinkedTreeMap<String, LinkedTreeMap<String, BigDecimal>> getRates() {
         return rates;
     }
 
-    public void setRates(LinkedTreeMap<String, LinkedTreeMap<String, Double>> rates) {
+    public void setRates(LinkedTreeMap<String, LinkedTreeMap<String, BigDecimal>> rates) {
         this.rates = rates;
     }
