+++ src/main/java/pl/model/Currency.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -22,13 +23,13 @@ public class Currency {
     @Column(name = "CUR_CURRENCY")
     private String currency;
     @Column(name = "CUR_VALUE")
-    private double value;
+    private BigDecimal value;
     //@Column(name = "CUR_VALUE2")
-    private LinkedTreeMap<String, Double> value2;
+    private LinkedTreeMap<String, BigDecimal> value2;
     /*@Column(name = "CUR_RATE")
     private LinkedTreeMap<String, Double> rates;*/
 
-    public Currency(String orderDate, String baseCurrency, String currency, Double value) {
+    public Currency(String orderDate, String baseCurrency, String currency, BigDecimal value) {
         this.orderDate = orderDate;
         this.baseCurrency = baseCurrency;
         this.currency = currency;
@@ -68,19 +69,19 @@ public class Currency {
         this.currency = currency;
     }
 
-    public double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
 
-    public LinkedTreeMap<String, Double> getValue2() {
+    public LinkedTreeMap<String, BigDecimal> getValue2() {
         return value2;
     }
 
-   public void setValue2(LinkedTreeMap<String, Double> value2){
+   public void setValue2(LinkedTreeMap<String, BigDecimal> value2){
             this.value2 = value2;
         }
 }
